Add tests for containers
